layering: playback service reads provider chrome from services, not widgets
Add lib/services/cloud/cloud_provider_presentation.dart with the pure
provider data (label, chip code, gradient, catalog chip/title, playlist
badge; dart:ui Color only). CloudProviderChrome (widgets) delegates to it
and keeps the Material icon and the bind-source chip. TorrentPlaybackService
imports the presentation instead of ../widgets/cloud_provider_chrome.dart;
the two widgets it hands an icon to (ProviderPickerOption,
showDebridActionSheet) now resolve the glyph from the provider id
themselves when no IconData is given, so the service passes none.
