+++ lib/widgets/cloud_provider_chrome.dart
@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
 
 import '../services/cloud/cloud_provider_id.dart';
+import '../services/cloud/cloud_provider_presentation.dart';
 import '../services/series_source_service.dart';
 
-/// Flutter colors/icons. String identity lives on [CloudProviderId].
+/// Flutter colors/icons. String identity lives on [CloudProviderId]; the
+/// label/code/gradient data lives on [CloudProviderPresentation] (services)
+/// and is re-exposed here so widget callers have one lookup.
 ///
 /// Lookup rules, on purpose:
 /// - [label]/[code]/[gradient]/[icon]: exact playback id (`debrid`, not `rd`)
@@ -13,60 +16,14 @@ import '../services/series_source_service.dart';
 class CloudProviderChrome {
   CloudProviderChrome._();
 
-  static const _indigo = [Color(0xFF6366F1), Color(0xFF4338CA)];
+  static String label(String provider) =>
+      CloudProviderPresentation.label(provider);
 
-  static String label(String provider) {
-    switch (provider) {
-      case 'preparing':
-        return 'Preparing';
-      case SeriesSource.localService:
-        return 'On-device';
-      case SeriesSource.addonDirectService:
-        return 'Direct addon';
-      case 'stream':
-        return 'Stream';
-      default:
-        return CloudProviderId.fromPlaybackId(provider)?.displayName ??
-            provider;
-    }
-  }
+  static String code(String provider) =>
+      CloudProviderPresentation.code(provider);
 
-  static String code(String provider) {
-    switch (provider) {
-      case 'preparing':
-        return '···';
-      case 'stream':
-        return 'TV';
-      case SeriesSource.addonDirectService:
-        return 'DL';
-      default:
-        final cloud = CloudProviderId.fromPlaybackId(provider);
-        if (cloud != null) return cloud.chipCode;
-        return provider.isEmpty ? '·' : provider.substring(0, 1).toUpperCase();
-    }
-  }
-
-  static List<Color> gradient(String provider) {
-    return switch (CloudProviderId.fromPlaybackId(provider)) {
-      CloudProviderId.debrid => const [
-        Color(0xFF10B981),
-        Color(0xFF059669),
-      ],
-      CloudProviderId.torbox => const [
-        Color(0xFF8B5CF6),
-        Color(0xFF7C3AED),
-      ],
-      CloudProviderId.premiumize => const [
-        Color(0xFFF59E0B),
-        Color(0xFFD97706),
-      ],
-      CloudProviderId.alldebrid => const [
-        Color(0xFF26A69A),
-        Color(0xFF00796B),
-      ],
-      CloudProviderId.pikpak || null => _indigo,
-    };
-  }
+  static List<Color> gradient(String provider) =>
+      CloudProviderPresentation.gradient(provider);
 
   static IconData icon(String provider) {
     return switch (CloudProviderId.fromPlaybackId(provider)) {
@@ -79,32 +36,17 @@ class CloudProviderChrome {
     };
   }
 
-  /// `auto` / unknown → `AUTO`, not the first letter of the string.
-  static String catalogChip(String provider) {
-    final id = CloudProviderId.tryParse(provider);
-    if (id == null) return 'AUTO';
-    return id.chipCode;
-  }
+  /// `auto` / unknown -> `AUTO`, not the first letter of the string.
+  static String catalogChip(String provider) =>
+      CloudProviderPresentation.catalogChip(provider);
 
-  static String? catalogTitle(String provider) {
-    return CloudProviderId.tryParse(provider)?.displayName;
-  }
+  static String? catalogTitle(String provider) =>
+      CloudProviderPresentation.catalogTitle(provider);
 
   /// Playlist card glyph. Not [catalogChip]: empty is RD, unknown is two
   /// letters, WebDAV is DV.
-  static String playlistBadge(String? raw) {
-    if (raw == null || raw.isEmpty) return CloudProviderId.debrid.chipCode;
-    switch (raw.toLowerCase()) {
-      case 'webdav':
-        return 'DV';
-      case 'pik-pak':
-      case 'pik_pak':
-        return CloudProviderId.pikpak.chipCode;
-      default:
-        return CloudProviderId.tryParse(raw)?.chipCode ??
-            raw.substring(0, 2).toUpperCase();
-    }
-  }
+  static String playlistBadge(String? raw) =>
+      CloudProviderPresentation.playlistBadge(raw);
 
   /// Bind-source chip. TorBox is blue here; playback [gradient] is purple.
   static ({String label, Color color}) sourceChip(String stored) => (
